+++ example/pinn/HeatEquationPINN_example.py
@@ -3,9 +3,9 @@ import torch
 import torch.nn as nn
 from torch.optim import Adam
 
-from fealpy.pinn.machine import LearningMachine
-from fealpy.pinn.boundary import TFC2dSpaceTimeDirichletBC
-from fealpy.pinn import gradient, ISampler
+from fealpy.ml.machine import LearningMachine
+from fealpy.ml.boundary import TFC2dSpaceTimeDirichletBC
+from fealpy.ml import gradient, ISampler
 
 # 定义神经网络
 pinn = nn.Sequential(
