+++ fealpy/ml/integral.py
