+++ fealpy/ml/machine.py
