+++ fealpy/ml/sampler/sampler.py
