+++ fealpy/ml/tools.py
